add config to resolvers
