+++ dtdl2-validator/dtdl2-validator.csproj.user
@@ -4,6 +4,7 @@
     <DebuggerFlavor>ProjectDebugger</DebuggerFlavor>
   </PropertyGroup>
   <PropertyGroup>
-    <ActiveDebugProfile>private-resolver</ActiveDebugProfile>
+    <ActiveDebugProfile>shared-resolver</ActiveDebugProfile>
+    <ShowAllFiles>true</ShowAllFiles>
   </PropertyGroup>
 </Project>
