+++ resolvers/LocalFolderResolver.cs
@@ -1,4 +1,7 @@
 ﻿using Microsoft.Azure.DigitalTwins.Parser;
+using Microsoft.Extensions.Configuration;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -8,15 +11,26 @@ namespace IoTModels.Resolvers
 {
     public class LocalFolderResolver : IResolver
     {
-        const string baseFolder = "models";
+        
+        string modelIndex;
         IDictionary<string, string> index = new Dictionary<string, string>();
-        public LocalFolderResolver()
+
+
+        public LocalFolderResolver(IConfiguration config)
         {
+            var baseFolder = config.GetValue<string>("baseFolder");
+            if (string.IsNullOrEmpty(baseFolder))
+            {
+                baseFolder = "models";
+            }
+
+            Console.WriteLine(new DirectoryInfo(baseFolder).FullName);
             if (!Directory.Exists(baseFolder))
             {
                 Console.WriteLine($"ERROR. BaseFolder '{baseFolder}' not found.");
             }
             TraverseDir(new DirectoryInfo(baseFolder));
+            
         }
 
         void TraverseDir(DirectoryInfo di)
@@ -25,7 +39,8 @@ namespace IoTModels.Resolvers
             {
                 foreach (var file in subfolder.GetFiles("*.json"))
                 {
-                    (string dtmi, string content) = LightParser.SemiParse(file.FullName);
+                    (string dtmi, string content) = SemiParse(file.FullName);
+                    // TODO: handle dependencies
                     index.Add(dtmi, content);
                     Console.WriteLine($"Found {dtmi} in local folder {subfolder.FullName}");
                 }
@@ -44,6 +59,36 @@ namespace IoTModels.Resolvers
                 }
             }
             return await Task.FromResult(resolvedModels);
+        }
+
+        (string dtmi, string content) SemiParse(string fileName)
+        {
+            if (!File.Exists(fileName))
+            {
+                throw new FileNotFoundException(fileName);
+            }
+            string content = File.ReadAllText(fileName);
+            JObject json;
+            try
+            {
+                json = JObject.Parse(content);
+            }
+            catch
+            {
+                throw new JsonReaderException("cant parse json");
+            }
+
+
+            var ctx = json.SelectToken("@context");
+            if (ctx.Value<string>() != "dtmi:dtdl:context;2")
+            {
+                throw new ApplicationException("Not valid DTDL v2");
+            }
+
+            var id = json.SelectToken("@id");
+
+
+            return (id.Value<string>(), content);
         }
     }
 }
